fully implemented shift limit screen

+++ MediaBazzar/MediaBazzar/ShiftLimitManager.cs
@@ -8,22 +8,22 @@ namespace MediaBazzar
 {
     public class ShiftLimitManager
     {
-        List<int> limits;
+        List<decimal> limits;
         ShiftLimitData data;
 
         public ShiftLimitManager()
         {
-            this.limits = new List<int>();
+            this.limits = new List<decimal>();
             this.data = new ShiftLimitData();
         }
 
-        public List<int> GetLimits()
+        public List<decimal> GetLimits()
         {
             limits = data.GetLimits();
             return limits;
         }
 
-        public void UpdateLimits(List<int> newLimits)
+        public void UpdateLimits(List<decimal> newLimits)
         {
             this.limits = newLimits;
             data.SetLimits(newLimits);
